データベース操作部をモジュール分けして整理
