+++ src-tauri/src/database/new.rs
@@ -1,0 +1,17 @@
+//! database構造体新規作成
+
+use super::*;
+
+impl Database {
+    /// 新規生成。
+    pub async fn new(host: &str, user: &str, pass: &str) -> Result<Self, DbError> {
+        let db_url = format!("mariadb://{}:{}@{}/nekotodo", user, pass, host);
+        let pool = MySqlPoolOptions::new()
+            .max_connections(10)
+            .min_connections(3)
+            .connect(&db_url)
+            .await
+            .map_err(DbError::FailConnect)?;
+        Ok(Self { pool })
+    }
+}
