+++ tex/list.tex
@@ -32,8 +32,28 @@
 \inputminted[linenos, breaklines]{rust}{src-rs/todo.rs}
 \clearpage
 
-\subsection{データベースアクセス database.rs}
+\subsection{データベースアクセスモジュール database.rs}
 \inputminted[linenos, breaklines]{rust}{src-rs/database.rs}
+\clearpage
+
+\subsubsection{新規作成 new.rs}
+\inputminted[linenos, breaklines]{rust}{src-rs/database/new.rs}
+\clearpage
+
+\subsubsection{セッション管理 session.rs}
+\inputminted[linenos, breaklines]{rust}{src-rs/database/session.rs}
+\clearpage
+
+\subsubsection{todoアイテム管理 todo.rs}
+\inputminted[linenos, breaklines]{rust}{src-rs/database/todo.rs}
+\clearpage
+
+\subsubsection{ユーザーデータ管理 user.rs}
+\inputminted[linenos, breaklines]{rust}{src-rs/database/user.rs}
+\clearpage
+
+\subsubsection{データベースモジュールテスト test.rs}
+\inputminted[linenos, breaklines]{rust}{src-rs/database/test.rs}
 \clearpage
 
 \subsection{commandモジュール tauri::command関数群 command.rs}
